Refactor statskeeper into its own package.
= Statskeeper is quite specific to be in a general utils package.
= This will allow statskeeper to access the probes/options package (package/options in turn depends on some methods in probeutils).

PiperOrigin-RevId: 278639167
